Refactor 'Remark' to 'Person' model
Remark has its own model ('model/remark/').

Enhance data encapsulation and align with the domain model.

Move Remark to Person model ('model/person/') directory.

Match the other fields of Person.
