+++ src/test/java/seedu/realodex/logic/parser/ParserUtilTest.java
@@ -26,7 +26,7 @@ import seedu.realodex.model.person.Family;
 import seedu.realodex.model.person.Income;
 import seedu.realodex.model.person.Name;
 import seedu.realodex.model.person.Phone;
-import seedu.realodex.model.remark.Remark;
+import seedu.realodex.model.person.Remark;
 import seedu.realodex.model.tag.Tag;
 
 public class ParserUtilTest {
