+++ src/test/java/seedu/realodex/testutil/PersonBuilder.java
@@ -10,7 +10,7 @@ import seedu.realodex.model.person.Income;
 import seedu.realodex.model.person.Name;
 import seedu.realodex.model.person.Person;
 import seedu.realodex.model.person.Phone;
-import seedu.realodex.model.remark.Remark;
+import seedu.realodex.model.person.Remark;
 import seedu.realodex.model.tag.Tag;
 import seedu.realodex.model.util.SampleDataUtil;
